Start running Appraisals in the build matrix
This goes further with splitting out different jobs, by listing the
Appraisals in the matrix. The hope is that this will provide better
feedback when parts of the build fails, as if we've got something like a
flakey test, it should run against every combination.

In addition to this, it extracts a "composite" action which allows for
some reuse between different jobs. It's not possible to reuse the
services section, unfortunately (YAML anchors aren't supported, which
would be the ideal solution).

https://docs.github.com/en/actions/using-jobs/using-a-matrix-for-your-jobs
https://docs.github.com/en/actions/creating-actions/creating-a-composite-action
https://github.com/actions/runner/issues/1182